requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
pathlib2==2.3.7
pygments==2.17.2
matplotlib==3.8.2
seaborn==0.13.1
gitpython==3.1.41
pyyaml==6.0.1
networkx==3.2.1
